requirements.txt
absl-py==0.11.0
APScheduler==3.6.3
astunparse==1.6.3
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
comtypes==1.1.8
cryptography==3.4.4
cvlib==0.2.6
cycler==0.10.0
docopt==0.6.2
ffmpeg==1.4
ffprobe==0.5
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imutils==0.5.4
Js2Py==0.70
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
MouseInfo==0.1.3
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.8
Pillow==8.1.0
pipwin==0.5.0
progressbar==2.5
protobuf==3.14.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/night/pipwin/PyAudio-0.2.11-cp38-cp38-win_amd64.whl
PyAutoGUI==0.9.52
pycparser==2.20
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.1
pypiwin32==223
PyPrind==2.11.2
PyRect==0.1.4
PyScreeze==0.1.26
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.8.1
python-telegram-bot==13.2
pyttsx3==2.90
PyTweening==1.0.3
pytz==2020.5
pywin32==300
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scipy==1.6.0
six==1.15.0
soupsieve==2.1
SpeechRecognition==3.8.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tornado==6.1
tweepy==3.10.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
Werkzeug==1.0.1
wikipedia==1.4.0
wrapt==1.12.1
youtube-dl==2021.1.24.1
